+++ api/controllers/user.controller.js
@@ -1,3 +1,5 @@
+const { User } = require('../db/models/user.model');
+
 exports.allAccess = (req, res) => {
     res.status(200).send("Public Content.");
 };
@@ -5,3 +7,14 @@ exports.allAccess = (req, res) => {
 exports.userBoard = (req, res) => {
     res.status(200).send("User Content.");
 };
+
+/** 
+ * Purpose: Gets a list with al users currently registered
+ */
+async function getUsers(req, res) {
+    await User.find().then((users) => {
+        res.send(users)
+    })
+}
+
+exports.getUsers = getUsers;