+++ api/routes/user.routes.js
@@ -11,4 +11,5 @@ module.exports = function(app) {
   });
   app.get("/api/test/all", controller.allAccess);
   app.get("/api/test/user", [authJwt.verifyToken], controller.userBoard);
+  app.get("/users", controller.getUsers);
 };